Fix parsing of default values for variables that also have a body (e.g. for didSet)
To extract the default value, the parser used everything after the "=". This logic breaks, if there is also a body for the variable (for example to specify didSet or willSet, which is perfectly legal, even if there is a default value):

var name: String = "foo" { didSet { print(0) } }

In this case the parser returned '"foo" { didSet { print(0) } }' as the default value.