Fix line number display of errors in swifttemplates
The line number for an unmatched tag was not computed properly, it was always
showing 0.
To fix this, we now keep a `processedComponents` array to keep track of the
template source that was already computed. This allows to know which line we are
at when we encounter an unmatched tag issue.