Remove clutter from releases
In zip versions, I noticed that some unrelated `Release` and `Sourcery.build`
folders were present.
This is because of the `:validate_docs` step which was writing them in the
`build/` during the process while the `:clean` step trashing `build/` had
already been done.
To prevent this, we now specify the derivedDataPath like in the `:build` step
and remove it after the process is done.