Create lenses only for stored properties (#525)
* Create lenses only for stored properties

Two changes, the first will create lenses only for stored properties. I guess they are not useful for computed properties?

The second change results in the type's initializer being only called with the stored properties, not the computed.

* Update CHANGELOG.md

* Update template tests for AutoLenses

No lenses for computed properties are generated and the initializer of Rectangle is only called with the stored properties.