Allow Swift templates (`.swifttemplate`) to define code blocks that have to be at the top level in the compiled `.swift` file.
Summary:
Currently, all code declarations from the processed `.swifttemplate` file are nested in the body of the function in the compiled `.swift` file like this:
```
extension TemplateContext {
    func generate() throws {
        \(contents)
    }
}
```

Such organization of the code prevents usage of some of the Swift language constructs that have to be at the top level of the Swift program, for example, protocol and type extensions.
The compiler would give "declaration is only valid at file scope" error if an extension was used in the template, for example.

However, when writing complex `.swifttemplate` templates it is sometimes desirable to be able to use them.

This commit adds support for a new escaping pattern `<%^` when parsing `.swifttemplate` files.
The code extracted from such blocks will be placed at the top level of the intermediate compiled `.swift` file, so that defining extensions becomes possible.

As an example, `Decorator.swifttemplate` is updated to define following extensions on `SourceryRuntime.Method` type instead of free functions:
```
fileprivate extension SourceryRuntime.Method {
    /// Constructs full method declaration string
    var methodDeclaration: String { /*...*/ }
    /// Constructs method call string passing in parameters with their local names
    var methodCall: String { /* ... */ )
}
```