Fix the creation of a symbolic link to bin/sourcery
Creating a symbolic link to `bin/sourcery` fails to execute as it will
resolve the path to Sourcery.app using the path of the link instead of
the path of the script file.
This fixes the path resolution using the code from this gist:
https://gist.github.com/TheMengzor/968e5ea87e99d9c41782
It will resolve the links until it finds the real path of the
`bin/sourcery` file.