Add support for protocol compositions
Partially fixes #541 (for named protocol compositions declared as a typealias) by adding a new `ProtocolComposition` `Type` subclass as described [here](https://github.com/krzysztofzablocki/Sourcery/issues/541#issuecomment-373544710).

The reason that this type is not named `TypeComposition` is that they are called "Protocol Compositions" in the [Swift reference manual](https://docs.swift.org/swift-book/ReferenceManual/Types.html#ID454).

A `ProtocolComposition` has an array of `composedTypes`, which reference each of the types composed between ampersands in its declaration. The `kind` of a `ProtocolComposition` is `"protocolComposition"`.

`ProtocolCompositions` are also made available on `Types` as `protocolCompositions`.