Fix group duplication while linking (#1)
Depending on the project setup, the children of the main group can have either path property, or name property, or both together.

For example, the Pods projects generated by Cocoapods would not add path property to the group, but only names.

Previously, sourcery was checking only path property to find out if the group exists or not. This PR adds check with the name of groups if there is no existing group with path.